+++ src/App.tsx
@@ -1,29 +1,16 @@
-import { ChangeEvent, useState } from 'react'
-import './App.css'
-import { Input } from './components/ui/input'
+import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+
+import { SearchCommander } from './pages/search-commander';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <SearchCommander />,
+  },
+]);
 
 function App() {
-  const [userInput, setUserInput] = useState<string>('')
-
-  const handleOnChange = (event: ChangeEvent<HTMLInputElement>) => {
-    setUserInput(event.target.value)
-  }
-
-  const fetchCommanders = async (name: string) => {
-    const response = await fetch(
-      `https://api.magicthegathering.io/v1/cards?name=${name}`
-    )
-    const resultsPromise = response.json()
-    const results = await Promise.resolve(resultsPromise)
-    console.log(results)
-  }
-
-  return (
-    <>
-      <Input value={userInput} onChange={handleOnChange} />
-      <button onClick={() => fetchCommanders(userInput)}>Search</button>
-    </>
-  )
+  return <RouterProvider router={router} />;
 }
 
-export default App
+export default App;
